+++ dist-esm/src/agent.js
@@ -243,4 +243,3 @@ export class HDCAgent {
         return agent;
     }
 }
-//# sourceMappingURL=agent.js.map